black style

+++ manim/renderer/opengl_renderer.py
@@ -446,9 +446,9 @@ class OpenGLRenderer:
 
     def get_frame(self):
         # get current pixel values as numpy data in order to test output
-        raw = self.get_raw_frame_buffer_object_data(dtype='f1')
+        raw = self.get_raw_frame_buffer_object_data(dtype="f1")
         result_dimensions = (config["pixel_height"], config["pixel_width"], 4)
-        np_buf = np.frombuffer(raw, dtype='uint8').reshape(result_dimensions)
+        np_buf = np.frombuffer(raw, dtype="uint8").reshape(result_dimensions)
         return np_buf
 
     # Returns offset from the bottom left corner in pixels.
